Initial fixture for git
Provide fixture object that can be used to build isolated test envs
when working with git and developing git tooling that uses the git
application.

Some basic tests on the topological sorting to ensure nodes are
returned in an expected order without requiring the code to enforce the
order.

Some simple acceptance tests ensure that the fixture will accept various
simple graphs and handle them correctly.

Change-Id: I5b39f8d52beca848adb20d646c26d91c4b977b92