Fix temp name creation and cleanup
Allow the system to delete the file immediately as once the name is
generated there is no further need of the temp file.

Fix the temp name splitting and generation to handle being provided a
string without the 'XXXXXX' template as part of it.

Temporarily disable the requirements.txt check as github3.py is not
part of global-requirements.

Change-Id: I14e56b0d4248dc037fab102acf3cb30573bfba8d